// Angular Inputs/FrontEnd/src/app/profile.interface.ts
export class initValue {
    name:string = "";
    email: string = "";
    password: string = "";
    mobile: string = "";
    rangeInput:number = 1;
    fav_game: string = "";
    gender: string = "";
    textArea: string = "";
    color: string = "#000000";
    background: string = "#FFFFFF";
    fileInput: string = "";
    condition: string = "";
}
export interface formdata {
    name: string;
    background: string
    email: string;
    password: string;
    mobile: string;
    rangeInput: number;
    fav_game: string;
    gender: string;
    textArea: string;
    color: string;
    fileInput: string;
    condition: string;
}